backend/requirements.txt
fastapi
onnxruntime
numpy
pillow
requests
beautifulsoup4
html5lib
lxml

